Add support for UserDrive with joysticks
When we started working with the stearing wheel and pedals we changed around the UserDrive
command so that it could be driven comfortably despite nearly everything being only buttons.
Unfortunately, we are likely not actually going to be driving with the wheel at competitions,
and as such we will have to use joysticks. Because of this I added a separate command which
allows the user to drive with joysticks, so we can switch back and forth as needed.