Update RobotMap. Change and Write commands and button assignments for robot function and testing to begin
The robot now can drive, climb, intake, and shoot from a software
perspective. Most of the control assignments are still on the
joystick, many of these assignments will be transferred to the xbox in a
future commit. Also expect the creation of default commands for both the
hopper and intake to utilize the xbox triggers.