Adjust commands, OI, shoot and hopper speeds. Begin auto template.
The shooter now succesfully adjusts its speed to a target encoder rate
value (currently 40). The default command to turn on the shooter and
hoppers from their correct buttons are written, but the hopper from the
xbox trigger is commented out from being set as the default command to
prevent interference with calling the command from the joystick
button(which is preferable in testing).