Fix camera conflicts, fix sequences, add current threshold to shooter.
The cameras were still having problems, but this is solved by ending both
cameras and both streams before starting either stream. The sequences for
Simult and SimultUltra (2 at once and 4 at once, respectively) were
tweaked to properly delay the pistons. The shooter has a blocking wall
that moves in response to the flywheel to prevent balls from entering from
the top of the hopper. If the shooter current spikes from having a ball
ground on it, all hopper pistons will open to hopefully clear the jam.