Prevent hopper pistons from opening when wheel is slow.
A problem at the reveal demo was that the code I wrote to open all the
hopper lanes when the shooter jammed was being triggered by the current
spike caused by the flywheel motors initialy powering on. To resolve this
I made sure a second had passed after enabling and the current was
stabilized before allowing that code to run. Additionaly, the
ShooterDefault will keep all the lanes closed if the wheel is below a
minimum speed in the first second of the wheel being turned on.