Move camera to 2014 and tune vision
The drive train directions reflect the wiring of 2014 which is being used
for testing. I tuned the HSL filter so that it does a much better job of
eliminating lights, and made the gear command run as a PI loop again (no
D). Future hope: add a gausian blur to the grip pipeline, just to make the
contours a little neater.